Testing a method for a test harness for examples.
The ADEVS examples are effectively tests and should be included as tests
in the build environments. But, they are also setup for user interaction
which is not useful for automated testing. This idea moves the main
functionality to a new header (<example_name>.h) which implements what
was originally in main.cpp. Both main.cpp and the new test.cpp files are
then wrappers around this that provide the correct input/output handling
that are expected for both tests and user examples. This may include
command line processing, argument handling, GUI setup, etc. Test.cpp
should use a default set of parameters that correspond to some expected
output.